% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts-for-the-times.R
\docType{data}
\name{tracts_for_the_times}
\alias{tracts_for_the_times}
\title{Tracts for the Times}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 3 columns.}
\usage{
tracts_for_the_times
}
\description{
Tracts for the Times
}
\keyword{datasets}

